model_deployment/requirements.txt
Flask==3.0.3
torch==2.3.1
torchvision==0.18.1
pillow==9.4.0